Improve contributors creation according to new specifications

+++ inc/cloner/class-cloner.php
@@ -1313,10 +1313,14 @@ class Cloner {
 		// Everything else
 		$book_information['pb_is_based_on'] = $this->sourceBookUrl;
 		$metadata_array_values = [ 'pb_keywords_tags', 'pb_bisac_subject', 'pb_additional_subjects' ];
+		$authors_slug = [];
 		foreach ( $book_information as $key => $value ) {
 			if ( $this->contributors->isValid( $key ) ) {
 				foreach ( $value as $contributor_data ) {
-					$this->contributors->insert( $contributor_data, $metadata_post_id, $key, $this->downloads );
+					$this->contributors->insert( $contributor_data, $metadata_post_id, $key, $this->downloads, 'slug' );
+					if ( $key === 'pb_authors' && isset( $contributor_data['slug'] ) ) {
+						$authors_slug[] = $contributor_data['slug'];
+					}
 				}
 			} elseif ( in_array( $key, $metadata_array_values, true ) ) {
 				$values = explode( ', ', $value );
@@ -1333,9 +1337,11 @@ class Cloner {
 			}
 		}
 
-		// Remove the current user from the author field in Book Info
 		$user_data = get_userdata( get_current_user_id() );
-		$this->contributors->unlink( $user_data->user_nicename, $metadata_post_id );
+		if ( ! in_array( $user_data->user_nicename, $authors_slug ) ) {
+			// Remove the current user from the author field in Book Info if it is not the author of the source book
+			$this->contributors->unlink( $user_data->user_nicename, $metadata_post_id );
+		}
 
 		return $metadata_post_id;
 	}
@@ -1616,7 +1622,7 @@ class Cloner {
 		foreach ( $section_information as $key => $value ) {
 			if ( $this->contributors->isValid( $key ) ) {
 				foreach ( $value as $contributor_data ) {
-					$this->contributors->insert( $contributor_data, $target_id, $key, $this->downloads );
+					$this->contributors->insert( $contributor_data, $target_id, $key, $this->downloads, 'slug' );
 				}
 			} else {
 				update_post_meta( $target_id, $key, $value );
